APM-3001 Refactor paas auth script
